@@ -12,112 +12,116 @@
     "url": "https://github.com/cypress-io/cypress-realworld-app/issues"
   },
   "engines": {
-    "node": ">=12"
+    "node": ">=12.22.0"
   },
   "dependencies": {
-    "@auth0/auth0-react": "1.8.0",
-    "@aws-amplify/ui-react": "1.2.8",
-    "@graphql-tools/graphql-file-loader": "7.1.0",
-    "@graphql-tools/load": "7.3.2",
-    "@material-ui/core": "4.12.3",
-    "@material-ui/icons": "4.11.2",
-    "@material-ui/lab": "4.0.0-alpha.60",
+    "@auth0/auth0-react": "1.9.0",
+    "@aws-amplify/ui-react": "1.2.26",
+    "@graphql-tools/graphql-file-loader": "7.3.10",
+    "@graphql-tools/load": "7.5.9",
+    "@material-ui/core": "4.12.4",
+    "@material-ui/icons": "4.11.3",
+    "@material-ui/lab": "4.0.0-alpha.61",
     "@okta/jwt-verifier": "2.3.0",
     "@okta/okta-auth-js": "4.9.2",
     "@okta/okta-react": "4.1.0",
-    "@xstate/react": "1.6.1",
-    "aws-amplify": "4.2.2",
-    "axios": "0.21.4",
+    "@types/detect-port": "^1.3.2",
+    "@xstate/react": "2.0.0",
+    "aws-amplify": "4.3.20",
+    "axios": "0.26.1",
     "clsx": "1.1.1",
-    "date-fns": "2.24.0",
-    "dinero.js": "1.9.0",
+    "date-fns": "2.28.0",
+    "detect-port": "1.3.0",
+    "dinero.js": "1.9.1",
     "formik": "2.2.9",
     "history": "4.10.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-google-login": "5.2.2",
     "react-infinite-calendar": "2.3.1",
-    "react-number-format": "4.7.3",
-    "react-router": "5.2.1",
-    "react-router-dom": "5.3.0",
+    "react-number-format": "4.9.1",
+    "react-router": "5.3.1",
+    "react-router-dom": "5.3.1",
     "react-virtualized": "9.22.3",
     "shortid": "2.2.16",
     "uuid": "8.3.2",
-    "xstate": "4.25.0",
-    "yup": "0.32.9"
+    "xstate": "4.31.0",
+    "yup": "0.32.11"
   },
   "devDependencies": {
-    "@cypress/code-coverage": "3.9.11",
+    "@cypress/code-coverage": "^3.10.0-dev.1",
     "@cypress/instrument-cra": "1.4.0",
+    "@cypress/react": "^5.10.0",
+    "@cypress/webpack-dev-server": "^1.6.0",
+    "@faker-js/faker": "6.1.2",
     "@percy/cypress": "2.3.4",
     "@types/bcryptjs": "2.4.2",
     "@types/bluebird": "3.5.36",
     "@types/connect-flash": "0.0.37",
     "@types/connect-history-api-fallback": "1.3.5",
     "@types/cors": "2.8.12",
-    "@types/dinero.js": "1.8.1",
+    "@types/dinero.js": "1.9.0",
     "@types/express": "4.17.2",
     "@types/express-paginate": "1.0.1",
     "@types/express-serve-static-core": "4.17.2",
     "@types/express-session": "1.17.4",
-    "@types/faker": "5.5.8",
+    "@types/faker": "5.5.9",
     "@types/http-proxy-middleware": "0.19.3",
-    "@types/jest": "27.0.1",
+    "@types/jest": "27.4.1",
     "@types/json-server": "0.14.4",
-    "@types/jsonwebtoken": "8.5.5",
+    "@types/jsonwebtoken": "8.5.8",
     "@types/jwt-decode": "2.2.1",
-    "@types/lodash": "4.14.172",
+    "@types/lodash": "4.14.181",
     "@types/lowdb": "1.0.11",
     "@types/morgan": "1.9.3",
-    "@types/node": "14.17.12",
+    "@types/node": "14.18.13",
     "@types/passport": "1.0.7",
-    "@types/react": "17.0.19",
-    "@types/react-dom": "17.0.9",
+    "@types/react": "17.0.44",
+    "@types/react-dom": "17.0.15",
     "@types/react-infinite-calendar": "2.3.6",
-    "@types/react-router": "5.1.16",
-    "@types/react-router-dom": "5.1.8",
-    "@types/react-virtualized": "9.21.13",
+    "@types/react-router": "5.1.18",
+    "@types/react-router-dom": "5.3.3",
+    "@types/react-virtualized": "9.21.21",
     "@types/shortid": "0.0.29",
-    "@types/uuid": "8.3.1",
-    "@types/validator": "13.6.3",
-    "@types/webpack-env": "1.16.2",
+    "@types/uuid": "8.3.4",
+    "@types/validator": "13.7.2",
+    "@types/webpack-env": "1.16.4",
     "@types/yup": "0.29.13",
-    "aws-sdk": "2.968.0",
+    "aws-sdk": "2.1116.0",
     "bcryptjs": "2.4.3",
-    "concurrently": "6.2.1",
+    "concurrently": "7.1.0",
     "cors": "2.8.5",
     "cross-env": "7.0.3",
-    "cypress": "^9.4.1",
-    "dotenv": "10.0.0",
-    "eslint-config-prettier": "8.3.0",
+    "cypress": "^10.1.0",
+    "dotenv": "16.0.0",
+    "eslint-config-prettier": "8.5.0",
     "eslint-plugin-cypress": "2.12.1",
     "eslint-plugin-prettier": "4.0.0",
-    "express": "4.17.1",
+    "express": "4.17.3",
     "express-graphql": "0.12.0",
-    "express-jwt": "6.1.0",
+    "express-jwt": "6.1.1",
     "express-paginate": "1.0.2",
     "express-session": "1.17.2",
-    "express-validator": "6.12.2",
-    "faker": "5.5.3",
-    "fuse.js": "6.4.6",
-    "graphql": "15.6.0",
-    "graphql-tools": "8.2.0",
-    "husky": "6.0.0",
-    "istanbul-lib-coverage": "3.0.1",
+    "express-validator": "6.14.0",
+    "fuse.js": "6.5.3",
+    "graphql": "16.4.0",
+    "graphql-tools": "8.2.7",
+    "husky": "7.0.4",
+    "istanbul-lib-coverage": "3.2.0",
     "json": "11.0.0",
-    "jwks-rsa": "2.0.4",
+    "jwks-rsa": "2.0.5",
     "lowdb": "1.0.0",
     "morgan": "1.10.0",
     "ncp": "2.0.0",
-    "nodemon": "2.0.13",
+    "nodemon": "2.0.15",
     "nyc": "15.1.0",
-    "passport": "0.4.1",
+    "passport": "0.5.2",
     "passport-local": "1.0.0",
-    "prettier": "2.4.1",
+    "prettier": "2.6.2",
     "react-scripts": "4.0.3",
     "start-server-and-test": "1.14.0",
-    "ts-node": "10.2.1",
-    "typescript": "4.4.2"
+    "ts-node": "10.7.0",
+    "typescript": "4.6.3"
   },
   "resolutions": {
     "@babel/compat-data": "7.9.0",
