@@ -3,7 +3,7 @@
 
 import * as jwt from "jsonwebtoken";
 
-Cypress.Commands.add("loginByAuth0Api", (username: string, password: string) => {
+Cypress.Commands.add("loginByAuth0Api", (username: string, password?: string) => {
   const log = Cypress.log({
     displayName: "AUTH0 LOGIN",
     message: [`🔐 Authenticating | ${username}`],
