@@ -9,7 +9,6 @@ import { isMobile } from "./utils";
 import "@percy/cypress";
 
 // Import commands for third-party auth providers
-import "./auth-provider-commands/cognito";
 import "./auth-provider-commands/auth0";
 import "./auth-provider-commands/okta";
 
@@ -342,7 +341,7 @@ Cypress.Commands.add("pickDateRange", (startDate, endDate) => {
     },
   });
 
-  const selectDate = (date: number) => {
+  const selectDate = (date: Date) => {
     return cy.get(`[data-date='${formatDate(date, "yyyy-MM-dd")}']`).click({ force: true });
   };
 
