@@ -1,4 +1,4 @@
-import faker from "faker";
+import { faker } from "@faker-js/faker";
 import { User } from "../../../src/models";
 
 const apiUsers = `${Cypress.env("apiUrl")}/users`;
