@@ -5,8 +5,7 @@ if (Cypress.env("auth0_client_id")) {
     beforeEach(function () {
       cy.task("db:seed");
 
-      cy.server();
-      cy.route("POST", "/bankAccounts").as("createBankAccount");
+      cy.route("POST", "/graphql").as("createBankAccount");
 
       cy.loginByAuth0Api(Cypress.env("auth0_username"), Cypress.env("auth0_password"));
     });
