@@ -1,16 +1,28 @@
 import React from "react";
 import { Snackbar } from "@material-ui/core";
-import { Interpreter } from "xstate";
+import {
+  BaseActionObject,
+  Interpreter,
+  ResolveTypegenMeta,
+  ServiceMap,
+  TypegenDisabled,
+} from "xstate";
 import { SnackbarContext, SnackbarSchema, SnackbarEvents } from "../machines/snackbarMachine";
-import { useService } from "@xstate/react";
+import { useActor } from "@xstate/react";
 import { Alert } from "@material-ui/lab";
 
 interface Props {
-  snackbarService: Interpreter<SnackbarContext, SnackbarSchema, SnackbarEvents, any>;
+  snackbarService: Interpreter<
+    SnackbarContext,
+    SnackbarSchema,
+    SnackbarEvents,
+    any,
+    ResolveTypegenMeta<TypegenDisabled, SnackbarEvents, BaseActionObject, ServiceMap>
+  >;
 }
 
 const AlertBar: React.FC<Props> = ({ snackbarService }) => {
-  const [snackbarState] = useService(snackbarService);
+  const [snackbarState] = useActor(snackbarService);
 
   return (
     <Snackbar
