@@ -1,7 +1,13 @@
 import React from "react";
 import clsx from "clsx";
-import { Interpreter } from "xstate";
-import { useService } from "@xstate/react";
+import {
+  BaseActionObject,
+  Interpreter,
+  ResolveTypegenMeta,
+  ServiceMap,
+  TypegenDisabled,
+} from "xstate";
+import { useActor } from "@xstate/react";
 import {
   makeStyles,
   AppBar,
@@ -21,7 +27,7 @@ import {
 } from "@material-ui/icons";
 import { Link as RouterLink, useLocation } from "react-router-dom";
 
-import { DataContext, DataEvents } from "../machines/dataMachine";
+import { DataContext, DataEvents, DataSchema } from "../machines/dataMachine";
 import TransactionNavTabs from "./TransactionNavTabs";
 import RWALogo from "./SvgRwaLogo";
 import RWALogoIcon from "./SvgRwaIconLogo";
@@ -80,14 +86,20 @@ const useStyles = makeStyles((theme) => ({
 interface NavBarProps {
   drawerOpen: boolean;
   toggleDrawer: Function;
-  notificationsService: Interpreter<DataContext, any, DataEvents, any>;
+  notificationsService: Interpreter<
+    DataContext,
+    DataSchema,
+    DataEvents,
+    any,
+    ResolveTypegenMeta<TypegenDisabled, DataEvents, BaseActionObject, ServiceMap>
+  >;
 }
 
 const NavBar: React.FC<NavBarProps> = ({ drawerOpen, toggleDrawer, notificationsService }) => {
   const match = useLocation();
   const classes = useStyles();
   const theme = useTheme();
-  const [notificationsState] = useService(notificationsService);
+  const [notificationsState] = useActor(notificationsService);
 
   const allNotifications = notificationsState?.context?.results;
   const xsBreakpoint = useMediaQuery(theme.breakpoints.only("xs"));
