@@ -1,5 +1,5 @@
 import React from "react";
-import { useService } from "@xstate/react";
+import { useActor } from "@xstate/react";
 import { Interpreter } from "xstate";
 import { Link } from "react-router-dom";
 import {
@@ -18,7 +18,7 @@ import { string, object, ref } from "yup";
 import RWALogo from "./SvgRwaLogo";
 import Footer from "./Footer";
 import { SignUpPayload } from "../models";
-import { AuthMachineContext, AuthMachineEvents } from "../machines/authMachine";
+import { AuthMachineContext, AuthMachineEvents, AuthMachineSchema } from "../machines/authMachine";
 
 const validationSchema = object({
   firstName: string().required("First Name is required"),
@@ -52,12 +52,12 @@ const useStyles = makeStyles((theme) => ({
 }));
 
 export interface Props {
-  authService: Interpreter<AuthMachineContext, any, AuthMachineEvents, any>;
+  authService: Interpreter<AuthMachineContext, AuthMachineSchema, AuthMachineEvents, any, any>;
 }
 
 const SignUpForm: React.FC<Props> = ({ authService }) => {
   const classes = useStyles();
-  const [, sendAuth] = useService(authService);
+  const [, sendAuth] = useActor(authService);
   const initialValues: SignUpPayload & { confirmPassword: string } = {
     firstName: "",
     lastName: "",
